% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{geo_gemeenten}
\alias{geo_ggdregios}
\alias{geo_nuts3}
\alias{geo_postcodes2}
\alias{geo_postcodes3}
\alias{geo_postcodes4}
\alias{geo_postcodes6}
\alias{geo_provincies}
\alias{postcodes}
\alias{postcodes4_afstanden}
\title{GIS Data Sets}
\usage{
geo_gemeenten

geo_ggdregios

geo_nuts3

geo_postcodes2

geo_postcodes3

geo_postcodes4

geo_postcodes6

geo_provincies

postcodes

postcodes4_afstanden
}
\description{
These data sets are inherited from the \code{certegis} package.
}
\seealso{
\itemize{
\item \link[certegis:cbs_geodata]{geo_gemeenten}
\item \link[certegis:cbs_geodata]{geo_ggdregios}
\item \link[certegis:cbs_geodata]{geo_nuts3}
\item \link[certegis:cbs_geodata]{geo_postcodes2}
\item \link[certegis:cbs_geodata]{geo_postcodes3}
\item \link[certegis:cbs_geodata]{geo_postcodes4}
\item \link[certegis:cbs_geodata]{geo_postcodes6}
\item \link[certegis:cbs_geodata]{geo_provincies}
\item \link[certegis:postcodes]{postcodes}
\item \link[certegis:postcodes4_afstanden]{postcodes4_afstanden}
}
}
\keyword{datasets}
