% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\docType{data}
\name{gis-datasets}
\alias{gis-datasets}
\alias{geo_gemeenten}
\alias{geo_ggdregios}
\alias{geo_nuts3}
\alias{geo_postcodes2}
\alias{geo_postcodes3}
\alias{geo_postcodes4}
\alias{geo_postcodes6}
\alias{geo_provincies}
\alias{postcodes}
\alias{postcodes4_afstanden}
\title{GIS Data Sets}
\usage{
geo_gemeenten

geo_ggdregios

geo_nuts3

geo_postcodes2

geo_postcodes3

geo_postcodes4

geo_postcodes6

geo_provincies

postcodes

postcodes4_afstanden
}
\description{
These data sets are inherited from the \code{certegis} package.
}
\examples{
geo_provincies

geo_provincies |>
  plot2(title = "Inwoners in Nederlandse provincies")

geo_provincies |>
  plot2(colour_fill = "umcgoranje")

geo_gemeenten |>
  filter_geolocation(provincie == "Groningen") |>
  plot2(category = inwoners / oppervlakte_km2,
        category.title = "Inwoners per km^2",
        title = "Populatiedichtheid in Groningse gemeenten")

geo_gemeenten |>
  filter_geolocation(provincie == "Groningen") |>
  plot2(category = inwoners / oppervlakte_km2,
        category.title = "Inwoners per km^2",
        title = "Populatiedichtheid in Groningse gemeenten",
        subtitle = "Met postcode-4 als overlay",
        datalabels = FALSE,
        colour = "umcgblauw",
        linewidth = 1) |>
  add_sf(geo_postcodes4 |> filter_geolocation(provincie == "Groningen"),
         colour = "umcgoranje",
         colour_fill = NA,
         linewidth = 0.25)
}
\seealso{
\itemize{
\item \link[certegis:cbs_geodata]{geo_gemeenten}
\item \link[certegis:cbs_geodata]{geo_ggdregios}
\item \link[certegis:cbs_geodata]{geo_nuts3}
\item \link[certegis:cbs_geodata]{geo_postcodes2}
\item \link[certegis:cbs_geodata]{geo_postcodes3}
\item \link[certegis:cbs_geodata]{geo_postcodes4}
\item \link[certegis:cbs_geodata]{geo_postcodes6}
\item \link[certegis:cbs_geodata]{geo_provincies}
\item \link[certegis:postcodes]{postcodes}
\item \link[certegis:postcodes4_afstanden]{postcodes4_afstanden}
}
}
\keyword{datasets}
