% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mmbi.epi-package}
\alias{mmbi.epi}
\alias{mmbi.epi-package}
\title{mmbi.epi: Helper Functions for Unit Epidemiology & Data Science of UMCG MMBI Department}
\description{
UMCG R package containing helper functions for functionalities used in their department of Medical Microbiology, unit Epidemiology & Data Science.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://umcg-mmbi-epidemiology.github.io/mmbi.epi/}
  \item \url{https://github.com/umcg-mmbi-epidemiology/mmbi.epi}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.s.berends@umcg.nl}

Authors:
\itemize{
  \item Gerolf C. de Boer
}

Other contributors:
\itemize{
  \item University Medical Center Groningen [copyright holder, funder]
}

}
\keyword{internal}
