% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scale_colour_umcg}
\alias{scale_colour_umcg}
\alias{scale_fill_umcg}
\alias{register_umcg_plot_style}
\alias{unregister_umcg_plot_style}
\title{UMCG Scale Colours}
\usage{
scale_colour_umcg(..., n)

scale_fill_umcg(..., n)

register_umcg_plot_style()

unregister_umcg_plot_style()
}
\arguments{
\item{...}{arguments passed on to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}} or \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}}

\item{n}{Number of colours required.}
}
\description{
These are \code{ggplot2} helpers to add organisational colours to \code{ggplot} objects.
}
\examples{
library(ggplot2)

p <- mtcars |>
  ggplot(aes(x = hp,
             y = mpg,
             colour = as.factor(vs)),
             fill = as.factor(vs)) +
  geom_point(size = 3) +
  geom_smooth()
p

p +
  plot_style_umcg(n = 2) +
  scale_fill_umcg(n = 2) +
  labs(title = "UMCG colours")

# With plot2(), this all goes automatically with less code
mtcars |>
  plot2(x = hp,
        y = mpg,
        category = as.factor(vs),
        smooth = TRUE,
        title = "Fully UMCG-styled plot",
        subtitle = "(with the right font too; 'Outfit' from Google Fonts)")
}
