% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R, R/plot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{add_map}
\alias{as.sf}
\alias{cases_within_radius}
\alias{convert_to_metre_CRS28992}
\alias{degrees_to_sf}
\alias{filter_geolocation}
\alias{filter_sf}
\alias{geocode}
\alias{is.sf}
\alias{latitude}
\alias{longitude}
\alias{reverse_geocode}
\alias{plot2}
\alias{get_colour}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{certegis}{\code{\link[certegis:GIS]{add_map}}, \code{\link[certegis:GIS]{as.sf}}, \code{\link[certegis]{cases_within_radius}}, \code{\link[certegis:GIS]{convert_to_metre_CRS28992}}, \code{\link[certegis:GIS]{degrees_to_sf}}, \code{\link[certegis:GIS]{filter_geolocation}}, \code{\link[certegis:GIS]{filter_sf}}, \code{\link[certegis:geocoding]{geocode}}, \code{\link[certegis:GIS]{is.sf}}, \code{\link[certegis:GIS]{latitude}}, \code{\link[certegis:GIS]{longitude}}, \code{\link[certegis:geocoding]{reverse_geocode}}}

  \item{plot2}{\code{\link[plot2:colour]{get_colour}}, \code{\link[plot2]{plot2}}}
}}

