% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scale_colour_umcg}
\alias{scale_colour_umcg}
\alias{scale_fill_umcg}
\title{UMCG Scale Colours}
\usage{
scale_colour_umcg(..., n)

scale_fill_umcg(..., n)
}
\arguments{
\item{n}{Number of colours required.}
}
\description{
UMCG Scale Colours
}
